+++ src/main/java/com/example/Taxi/domain/Room.java
@@ -18,11 +18,11 @@ public class Room {
     private Long id;
     private String destination;
     private LocalDateTime dateTime;
-    private BigDecimal latitude; //위도
-    private BigDecimal longitude; //경도
+    private Double latitude; //위도
+    private Double longitude; //경도
 
     @Builder
-    public Room(String destination, LocalDateTime dateTime, BigDecimal latitude, BigDecimal longitude) {
+    public Room(String destination, LocalDateTime dateTime, Double latitude, Double longitude) {
         this.destination = destination;
         this.dateTime = dateTime;
         this.latitude = latitude;
